Rename preact10-src to preact10-shallow-diff
