+++ src/compat-shallow-renderer/preact10-shallow-diff.ts
